add docs and polish
